refactor: Changing and separating content app
