+++ StreamSync/settings.py
@@ -36,7 +36,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'users'
+    'apps.users',
+    'apps.analytics',
+    'apps.contents',
+    'apps.lists'
 ]
 
 MIDDLEWARE = [
