+++ StreamSync/urls.py
@@ -2,11 +2,14 @@
 from django.contrib import admin
 from django.urls import path, include
 from django.contrib.auth import views as auth_views
-from users.views import StreamSyncLoginView
+from apps.users.views import StreamSyncLoginView
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('users.urls')),
+    path('', include('apps.users.urls')),
+    path('', include('apps.contents.urls')),
+    # path('', include('apps.analytics.urls')),
+    # path('', include('apps.lists.urls')),
 
     path('login/', StreamSyncLoginView.as_view(template_name='registration/login.html'), name='login'),
