+++ apps/analytics/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class AnalyticsConfig(AppConfig):
-    name = 'analytics'
+    name = 'apps.analytics'
