+++ apps/contents/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class ContentsConfig(AppConfig):
-    name = 'contents'
+    name = 'apps.contents'
