+++ apps/external_apis.py
@@ -1,0 +1,38 @@
+import os
+
+import requests
+from dotenv import load_dotenv
+
+# 1. CARREGUEM CONFIGURACIÓ
+load_dotenv()
+
+urls_list = os.getenv('API_BASE_URLS', '').split(',')
+keys_list = os.getenv('API_KEYS_DJANGO', '').split(',')
+API_CONFIG = dict(zip(urls_list, keys_list))
+
+
+def get_genres_from_api():
+    for base_url, key in API_CONFIG.items():
+        try:
+            return requests.get(f"{base_url}/genres", headers={'x-api-key': key}, timeout=1).json()
+        except:
+            continue
+    return []
+
+
+def get_directors_from_api():
+    for base_url, key in API_CONFIG.items():
+        try:
+            return requests.get(f"{base_url}/directors", headers={'x-api-key': key}, timeout=1).json()
+        except:
+            continue
+    return []
+
+
+def get_age_ratings_from_api():
+    for base_url, key in API_CONFIG.items():
+        try:
+            return requests.get(f"{base_url}/age-ratings", headers={'x-api-key': key}, timeout=1).json()
+        except:
+            continue
+    return []
