+++ apps/lists/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class ListsConfig(AppConfig):
-    name = 'lists'
+    name = 'apps.lists'
