+++ apps/users/views.py
@@ -1,37 +1,30 @@
 import os
 import requests
-from concurrent.futures import ThreadPoolExecutor
 from datetime import timedelta
-from django.core.paginator import Paginator
-from django.http import HttpResponse
 from django.http import JsonResponse
-from django.shortcuts import render, redirect, get_object_or_404
+from django.shortcuts import render, get_object_or_404
 from django.contrib.auth import update_session_auth_hash, login
 from django.contrib.auth.views import LoginView
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.contrib.auth.forms import PasswordChangeForm
-from dotenv import load_dotenv
-from rest_framework.decorators import api_view
-from thefuzz import process, fuzz
 from django.db.models import Count, Avg, Sum
 from django.utils import timezone
 
-from .models import Pelicula, LlistaPersonal, Carpeta, Profile, Ressenya, Views, Feedback
-from .forms import UserRegistrationForm, UserUpdateForm
+from apps.external_apis import get_genres_from_api, get_age_ratings_from_api
+
+from apps.users.decorators.permissions import cap_manager_permes
+from apps.users.models.models import  LlistaPersonal, Carpeta, Profile, Ressenya, Views, Feedback
+from apps.users.forms.forms import UserRegistrationForm, UserUpdateForm
 
 # 1. LOAD CONFIGURATION
-from functools import wraps
 from django.shortcuts import redirect
 import json
-# 1. CARREGUEM CONFIGURACIÓ
-load_dotenv()
 
-urls_list = os.getenv('API_BASE_URLS', '').split(',')
-keys_list = os.getenv('API_KEYS_DJANGO', '').split(',')
-API_CONFIG = dict(zip(urls_list, keys_list))
+from apps.contents.models import Pelicula
+from apps.contents.service import get_all_series, get_all_movies, enrich_tmdb_images
 
-TMDB_API_KEY = os.getenv('TMDB_API_KEY')
+
 
 OPTIONS = {
     'plataformas': ['CinePlus', 'StreamHub', 'PlayMax'],
@@ -54,114 +47,6 @@ class StreamSyncLoginView(LoginView):
         messages.success(self.request, f"Benvingut/da de nou, {form.get_user().username}!")
         return response
 
-# --- decorador usuari SPM---
-
-def cap_manager_permes(view_func):
-
-    @wraps(view_func)
-    def _wrapped_view(request, *args, **kwargs):
-        if request.user.is_authenticated and hasattr(request.user, 'profile'):
-            plataforma = request.user.profile.manager_de
-            if plataforma:
-                # Es un SPM. Lo mandamos a su panel.
-                messages.info(request, "Ets un Manager. Aquesta és la teva àrea de treball.")
-                return redirect('dashboard_manager', plataforma_nom=plataforma)
-        return view_func(request, *args, **kwargs)
-    return _wrapped_view
-# --- 2. FUNCIONS AUXILIARS I MAPEIG ---
-
-# --- 2. TMDB FUNCTIONS ---
-
-def get_tmdb_image(title):
-    try:
-        response = requests.get(
-            "https://api.themoviedb.org/3/search/multi",
-            params={
-                "api_key": TMDB_API_KEY,
-                "query": title,
-                "language": "en"
-            },
-            timeout=2
-        )
-        if response.status_code == 200:
-            results = response.json().get("results", [])
-            if results and results[0].get("poster_path"):
-                return f"https://image.tmdb.org/t/p/w500{results[0]['poster_path']}"
-    except (requests.RequestException, ValueError):
-        pass
-    return 'https://via.placeholder.com/300x450'
-
-
-def enrich_tmdb_images(items):
-    def load_image(item):
-        item['imatge'] = get_tmdb_image(item['titol'])
-        return item
-
-    with ThreadPoolExecutor(max_workers=10) as executor:
-        items = list(executor.map(load_image, items))
-    return items
-
-
-# --- 3. DATA MAPPING ---
-
-def map_data(item, port):
-    platforms = {"8080": "CinePlus", "8081": "StreamHub", "8082": "PlayMax"}
-    port_net = str(port).replace('/','')
-
-
-    title = item.get('title') or item.get('titol') or "Sense títol"
-    synopsis = item.get('synopsis') or "Sense sinopsi disponible."
-    content_year = item.get('year') or item.get('start_year') or 0
-
-    return {
-        'id': f"{port}_{item.get('id')}",
-        'titol': title,
-        'sinopsi': synopsis,
-        'any': content_year,
-        'any_fi': item.get('end_year'),
-        'total_seasons': item.get('total_seasons'),
-        'rating': item.get('rating', '0.0'),
-        'imatge': item.get('imatge') or 'https://via.placeholder.com/300x450',
-        'plataforma': platforms.get(port, "Altres"),
-        'genre_id': item.get('genre_id'),
-        'director_id': item.get('director_id'),
-        'age_rating_id': item.get('age_rating_id'),
-        'genere_nom': "General",
-        'director_nom': "Desconegut",
-        'edat_nom': "N/A"
-    }
-
-def deduplicate_content(llista):
-    vistos = {}
-    for item in llista:
-        titol = item['titol'].strip().lower()
-        if titol not in vistos:
-            item['plataformes_disponibles'] = [item['plataforma']]
-            vistos[titol] = item
-        else:
-            if item['plataforma'] not in vistos[titol]['plataformes_disponibles']:
-                vistos[titol]['plataformes_disponibles'].append(item['plataforma'])
-    return list(vistos.values())
-
-# --- 4. STREAMSYNC API CALLS ---
-
-def get_all_movies(query=None):
-    results = []
-    for base_url, key in API_CONFIG.items():
-        headers = {'x-api-key': key}
-        port = base_url.split(':')[-1]
-        params = {'title': query} if query else {}
-        try:
-            response = requests.get(f"{base_url}/movies", headers=headers, params=params, timeout=2)
-            if response.status_code == 200:
-                for item in response.json():
-                    obj = map_data(item, port)
-                    obj['tipus'] = 'movie'
-                    results.append(obj)
-        except:
-            pass
-    return deduplicate_content(results)  # ✅ Deduplicació
-
 
 def enrich_api_data(content_list):
     genres_api = get_genres_from_api()
@@ -179,51 +64,6 @@ def enrich_api_data(content_list):
             item['tipus'] = item.get('media_type', 'movie')
 
     return content_list
-
-
-def get_all_series(query=None):
-    results = []
-    for base_url, key in API_CONFIG.items():
-        headers = {'x-api-key': key}
-        port = base_url.split(':')[-1]
-        params = {'title': query} if query else {}
-        try:
-            response = requests.get(f"{base_url}/series", headers=headers, params=params, timeout=2)
-            if response.status_code == 200:
-                for item in response.json():
-                    obj = map_data(item, port)
-                    obj['tipus'] = 'series'
-                    results.append(obj)
-        except:
-            pass
-    return deduplicate_content(results)  # ✅ Deduplicació
-
-
-def get_genres_from_api():
-    for base_url, key in API_CONFIG.items():
-        try:
-            return requests.get(f"{base_url}/genres", headers={'x-api-key': key}, timeout=1).json()
-        except:
-            continue
-    return []
-
-
-def get_directors_from_api():
-    for base_url, key in API_CONFIG.items():
-        try:
-            return requests.get(f"{base_url}/directors", headers={'x-api-key': key}, timeout=1).json()
-        except:
-            continue
-    return []
-
-
-def get_age_ratings_from_api():
-    for base_url, key in API_CONFIG.items():
-        try:
-            return requests.get(f"{base_url}/age-ratings", headers={'x-api-key': key}, timeout=1).json()
-        except:
-            continue
-    return []
 
 
 # --- 4. MAIN VIEWS ---
@@ -289,137 +129,8 @@ def home_page(request):
     })
 
 
-@cap_manager_permes
-def content_detail(request, tipus, content_id):
-    all_content = get_all_series() if tipus == 'series' else get_all_movies()
-    item = next((p for p in all_content if str(p['id']) == str(content_id)), None)
 
-    if not item:
-        return render(request, '404.html', status=404)
 
-    if 'plataformes_disponibles' not in item:
-        item['plataformes_disponibles'] = [item.get('plataforma', 'N/A')]
-
-    # ✅ Carreguem traduccions una sola vegada
-    genres = get_genres_from_api()
-    directors = get_directors_from_api()
-    ratings = get_age_ratings_from_api()
-
-    mapa_genres = {str(g['id']): g['name'] for g in genres}
-    mapa_ratings = {str(r['id']): r.get('description') or r.get('title') or r.get('name') or 'N/A' for r in ratings}
-    mapa_directors = {str(d['id']): d['name'] for d in directors}
-
-    # ✅ Enriquim l'item principal
-    item['genere_nom'] = mapa_genres.get(str(item.get('genre_id')), 'General')
-    item['director_nom'] = mapa_directors.get(str(item.get('director_id')), 'Desconegut')
-    item['edat_nom'] = mapa_ratings.get(str(item.get('age_rating_id')), 'N/A')
-    item['imatge'] = get_tmdb_image(item['titol'])
-
-    movie_db, _ = Pelicula.objects.update_or_create(
-        id=item['id'],
-        defaults={
-            "titol": item['titol'],
-            "any": item['any'],
-            "valoracio": float(item.get('rating', 0)),
-            "imatge": item.get('imatge'),
-            "tipus": tipus,
-            "plataforma": item.get('plataforma')
-        }
-    )
-
-    # ✅ Enriquim les recomanacions amb gènere i edat
-    raw_recommendations = [p for p in all_content if str(p['id']) != str(content_id)][:5]
-    for r in raw_recommendations:
-        r['genere_nom'] = mapa_genres.get(str(r.get('genre_id')), 'General')
-        r['edat_nom'] = mapa_ratings.get(str(r.get('age_rating_id')), 'N/A')
-
-    recommendations = enrich_tmdb_images(raw_recommendations)
-
-    # CHECK IF THERE'S ALREADY A REVIEW
-    ressenya_usuari = None
-
-    if request.user.is_authenticated:
-        ressenya_usuari = Ressenya.objects.filter(
-            usuari=request.user,
-            pelicula=movie_db
-        ).first()
-
-    return render(request, 'pagina_contingut.html', {
-        'item': item,
-        'tipus': tipus,
-        'ja_guardada': LlistaPersonal.objects.filter(usuari=request.user,
-                                                     pelicula=movie_db).exists() if request.user.is_authenticated else False,
-        'carpetes': request.user.les_meves_carpetes.all() if request.user.is_authenticated else [],
-        'ressenyes': Ressenya.objects.filter(pelicula=movie_db).order_by('-data_publicacio'),
-        'recomanacions': recommendations,
-        'ressenya_usuari': ressenya_usuari,
-    })
-
-@cap_manager_permes
-def catalogo(request, tipus=None):
-    if tipus == 'movie':
-        all_content = get_all_movies()
-    elif tipus == 'series':
-        all_content = get_all_series()
-    else:
-        all_content = get_all_movies() + get_all_series()
-
-    genres_api = get_genres_from_api()
-    ratings_api = get_age_ratings_from_api()
-    directors_api = get_directors_from_api()
-
-    genre_map = {str(g['id']): g['name'] for g in genres_api}
-    rating_map = {str(r['id']): r.get('description', 'N/A') for r in ratings_api}
-    director_map = {str(d['id']): d['name'] for d in directors_api}
-
-    filters = {
-        'p': request.GET.get('plataforma', ''),
-        'g': request.GET.get('genere', ''),
-        'e': request.GET.get('edat', ''),
-        'v': request.GET.get('valoracio', '0'),
-        'd': request.GET.get('director', '').strip().lower()
-    }
-
-    results = []
-    for item in all_content:
-        gid = str(item.get('genre_id'))
-        eid = str(item.get('age_rating_id'))
-        did = str(item.get('director_id'))
-
-        item['genere_nom'] = genre_map.get(gid, "General")
-        item['edat_nom'] = rating_map.get(eid, "N/A")
-        item['director_nom'] = director_map.get(did, "Desconegut")
-
-        if filters['p'] and filters['p'] not in item.get('plataformes_disponibles', [item.get('plataforma', '')]): continue
-        if filters['g'] and gid != filters['g']: continue
-        if filters['e'] and eid != filters['e']: continue
-        if filters['d'] and filters['d'] not in item['director_nom'].lower(): continue
-
-        try:
-            if float(item.get('rating', 0)) < float(filters['v']): continue
-        except:
-            pass
-
-        results.append(item)
-
-    paginator = Paginator(results, 12)
-    page_number = request.GET.get('page', 1)
-    page_obj = paginator.get_page(page_number)
-
-    enrich_tmdb_images(list(page_obj.object_list))
-
-    filters_url = f"&plataforma={filters['p']}&genere={filters['g']}&edat={filters['e']}&valoracio={filters['v']}&director={filters['d']}"
-
-    return render(request, 'cataleg.html', {
-        'contenidos': page_obj.object_list,
-        'page_obj': page_obj,
-        'filtres_url': filters_url,
-        'tipus_actual': tipus,
-        'opcions': OPTIONS,
-        'genres_api': genres_api,
-        'ratings': ratings_api,
-        'filtros_sel': filters
-    })
 
 def feedback_view(request):
     if request.method == "POST":
@@ -473,13 +184,14 @@ def delete_review(request, ressenya_id):
     return redirect('pagina_contingut', tipus=content_type, content_id=content_id_value)
 
 
-@login_required
-def lists(request):
-    pass
+# @login_required
+# def lists(request):
+#     pass
 
 
 @cap_manager_permes
-def llistes(request):
+def lists(request):
+    print(request.user)
     return render(request, 'llistes.html', {
         'carpetes': request.user.les_meves_carpetes.all(),
         'elements_solts': LlistaPersonal.objects.filter(usuari=request.user, carpeta__isnull=True)
@@ -629,69 +341,7 @@ def delete_account(request):
     return render(request, 'registration/esborrar_compte.html')
 
 
-def search_content(request):
-    query = request.GET.get('q', '').strip()
 
-    movies = get_all_movies()
-    for m in movies: m['tipus'] = 'movie'
-    series = get_all_series()
-    for s in series: s['tipus'] = 'series'
-
-    all_content = movies + series
-    main_result = None
-    recommendations = []
-
-    # ✅ Carreguem traduccions una sola vegada
-    genres_api = get_genres_from_api()
-    ratings_api = get_age_ratings_from_api()
-    mapa_genres = {str(g['id']): g['name'] for g in genres_api}
-    mapa_ratings = {str(r['id']): r.get('description', 'N/A') for r in ratings_api}
-
-    if query:
-        titles = [p['titol'] for p in all_content]
-        matches = process.extract(query, titles, scorer=fuzz.token_set_ratio, limit=1)
-
-        if matches and matches[0][1] > 65:
-            main_result = next(p for p in all_content if p['titol'] == matches[0][0])
-
-            # ✅ Enriquim el resultat principal
-            main_result['genere_nom'] = mapa_genres.get(str(main_result.get('genre_id')), 'General')
-            main_result['edat_nom'] = mapa_ratings.get(str(main_result.get('age_rating_id')), 'N/A')
-            main_result['imatge'] = get_tmdb_image(main_result['titol'])
-
-            others = [p for p in all_content if p['id'] != main_result['id']]
-
-            def calculate_score(item):
-                score = 0
-                if item.get('director_id') == main_result.get('director_id'):
-                    score += 10
-                if item.get('genre_id') == main_result.get('genre_id'):
-                    score += 5
-                if item.get('age_rating_id') == main_result.get('age_rating_id'):
-                    score += 2
-                return score
-             
-            
-
-            # ✅ Només continguts del mateix gènere
-            recommendations = [
-                p for p in others
-                if p.get('genre_id') == main_result.get('genre_id')
-            ][:5]
-
-            # ✅ Enriquim gènere i edat de les recomanacions
-            for item in recommendations:
-                item['genere_nom'] = mapa_genres.get(str(item.get('genre_id')), 'General')
-                item['edat_nom'] = mapa_ratings.get(str(item.get('age_rating_id')), 'N/A')
-
-
-            enrich_tmdb_images(recommendations)
-
-    return render(request, 'cerca_contingut.html', {
-        'query': query,
-        'resultat': main_result,
-        'resultats': recommendations
-    })
 
 
 @login_required
